dev work

+++ tokenizer.c
@@ -250,7 +250,7 @@ void tokenizer_next(void) {
   return;
 }
 /*---------------------------------------------------------------------------*/
-VARIABLE_TYPE tokenizer_num(void) { return atoi(ptr); }
+VARIABLE_TYPE tokenizer_num(void) { return atol(ptr); }
 /*---------------------------------------------------------------------------*/
 VARFLOAT_TYPE tokenizer_numfloat(void) { return atof(ptr); }
 /*---------------------------------------------------------------------------*/
